pyproject.toml
[project]
name = "goidama"
version = "0.1.0"
description = ""
authors = [
    {name = "myeongsu",email = "puggiboy0304@naver.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.136.0,<0.137.0)",
    "uvicorn (>=0.44.0,<0.45.0)",
    "sqlmodel (>=0.0.38,<0.0.39)",
    "aiosqlite (>=0.22.1,<0.23.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "greenlet (>=3.4.0,<4.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "pydantic-settings (>=2.14.0,<3.0.0)",
    "pyjwt (>=2.12.1,<3.0.0)",
    "aioboto3 (>=15.5.0,<16.0.0)",
    "pillow (>=12.2.0,<13.0.0)",
    "geopy (>=2.4.1,<3.0.0)",
    "python-multipart (>=0.0.26,<0.0.27)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "apscheduler (>=3.11.2,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.3,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)"
]

[tool.pytest.ini_options]
minversion = "8.3.4"
env = []
addopts = " --strict-markers --tb=short --asyncio-mode=auto -p no:warnings --doctest-modules"
python_files = ["tests.py", "test_*.py"]
rootdir = "./"
testpaths = ["./tests", "./appserver"]
pythonpath = ["./"]
filterwarnings = ["error", "ignore::DeprecationWarning:etcd3.*:"]
log_cli = true
log_cli_level = "WARNING"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"